# ra-idp/gen/openapi.yaml
openapi: 3.0.3
info:
  title: "RA OAuth2 / OIDC IdP"
  version: "1.0"
paths:
...
  /authorize:
    get:
      operationId: authorize
      summary: Authorization endpoint (RFC 6749 §4.1.1)
      responses:
        302:
          description: Authorization endpoint (RFC 6749 §4.1.1)
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: scope
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: nonce
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum:
              - S256
        - name: prompt
          in: query
          required: false
          schema:
            type: string
        - name: max_age
          in: query
          required: false
          schema:
            type: integer
        - name: id_token_hint
          in: query
          required: false
          schema:
            type: string
        - name: request_uri
          in: query
          required: false
          schema:
            type: string
...
  /end_session:
    get:
      operationId: endSession
      summary: "RP-Initiated Logout endpoint (OIDC RP-Initiated Logout 1.0)。redirect 検証は OAuth2/OIDC component が行い、session revocation は Authentication component の SessionManager に委譲する。"
      responses:
        200:
          description: "RP-Initiated Logout endpoint (OIDC RP-Initiated Logout 1.0)。redirect 検証は OAuth2/OIDC component が行い、session revocation は Authentication component の SessionManager に委譲する。"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
      parameters:
        - name: client_id
          in: query
          required: false
          schema:
            type: string
        - name: id_token_hint
          in: query
          required: false
          schema:
            type: string
        - name: post_logout_redirect_uri
          in: query
          required: false
          schema:
            type: string
            format: uri
        - name: state
          in: query
          required: false
          schema:
            type: string
...
  /introspect:
    post:
      operationId: introspect
      summary: Token Introspection (RFC 7662)
      responses:
        200:
          description: Token Introspection (RFC 7662)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrospectionResponse"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/IntrospectionRequest"
  /revoke:
    post:
      operationId: revoke
      summary: Token Revocation (RFC 7009)
      responses:
        200:
          description: Token Revocation (RFC 7009)
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
...
components:
  schemas:
...
    SenderConstraintCnf:
      type: object
      additionalProperties: false
      description: "Sender-Constrained 証明 (RFC 7800)。"
      properties:
        jkt:
          type: string
        x5t_s256:
          type: string
...
    IntrospectionRequest:
      type: object
      additionalProperties: false
      description: /introspect リクエスト (RFC 7662 §2.1)。ResourceServer がトークンの有効性を AuthorizationServer に問い合わせる。
      properties:
        token:
          type: string
        token_type_hint:
          type: string
          enum:
            - access_token
            - refresh_token
      required:
        - token
    IntrospectionResponse:
      type: object
      additionalProperties: false
      description: /introspect レスポンス (RFC 7662 §2.2)。トークンの active 状態と有効時のメタデータを返す。
      properties:
        active:
          type: boolean
        scope:
          type: string
        client_id:
          type: string
        sub:
          type: string
        exp:
          type: integer
        iat:
          type: integer
        token_type:
          type: string
        jti:
          type: string
        cnf:
          $ref: "#/components/schemas/SenderConstraintCnf"
      required:
        - active
    RevokeRequest:
      type: object
      additionalProperties: false
      description: /revoke リクエスト (RFC 7009 §2.1)。指定したトークンを失効させる。
      properties:
        token:
          type: string
        token_type_hint:
          type: string
          enum:
            - access_token
            - refresh_token
      required:
        - token
...
    OAuthError:
      type: object
      properties:
        type:
          const: OAuthError
          type: string
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - invalid_token
            - server_error
            - invalid_dpop_proof
            - access_denied
            - expired_token
            - slow_down
            - insufficient_scope
            - authorization_pending
        error_description:
          type: string
        error_uri:
          type: string
      required:
        - type
        - error
...